# templates/ts-web/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KDP Text Tool",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/api/v1/context": {
      "get": {
        "operationId": "getContext",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "userId",
                        "toolId",
                        "displayName",
                        "toolName",
                        "description",
                        "environment",
                        "portalUrl"
                      ],
                      "properties": {
                        "userId": {
                          "type": "string"
                        },
                        "toolId": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "toolName": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "environment": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "development"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "production"
                              ]
                            }
                          ]
                        },
                        "portalUrl": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/normalize": {
      "post": {
        "operationId": "normalizeText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "text",
                  "options"
                ],
                "properties": {
                  "text": {
                    "minLength": 1,
                    "maxLength": 100000,
                    "type": "string"
                  },
                  "options": {
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "trimLines",
                      "collapseSpaces",
                      "removeEmptyLines"
                    ],
                    "properties": {
                      "trimLines": {
                        "type": "boolean"
                      },
                      "collapseSpaces": {
                        "type": "boolean"
                      },
                      "removeEmptyLines": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "text",
                        "inputCharacters",
                        "outputCharacters",
                        "inputLines",
                        "outputLines",
                        "removedEmptyLines"
                      ],
                      "properties": {
                        "text": {
                          "type": "string"
                        },
                        "inputCharacters": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "outputCharacters": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "inputLines": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "outputLines": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "removedEmptyLines": {
                          "minimum": 0,
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}